qa: abstract the signer from the Lianad class
That's prep work to introduce a multisig
