+++ tests/fixtures.py
@@ -3,6 +3,7 @@ from bip380.descriptors import Descriptor
 from concurrent import futures
 from test_framework.bitcoind import Bitcoind
 from test_framework.lianad import Lianad
+from test_framework.signer import SingleSigner
 from test_framework.utils import (
     EXECUTOR_WORKERS,
 )
@@ -118,18 +119,19 @@ def lianad(bitcoind, directory):
     os.makedirs(datadir, exist_ok=True)
     bitcoind_cookie = os.path.join(bitcoind.bitcoin_dir, "regtest", ".cookie")
 
-    owner_hd = BIP32.from_seed(os.urandom(32), network="test")
-    recovery_hd = BIP32.from_seed(os.urandom(32), network="test")
-    owner_xpub, recovery_xpub = owner_hd.get_xpub(), recovery_hd.get_xpub()
+    signer = SingleSigner()
+    primary_xpub, recovery_xpub = (
+        signer.primary_hd.get_xpub(),
+        signer.recovery_hd.get_xpub(),
+    )
     csv_value = 10
     main_desc = Descriptor.from_str(
-        f"wsh(or_d(pk({owner_xpub}/<0;1>/*),and_v(v:pkh({recovery_xpub}/<0;1>/*),older({csv_value}))))"
+        f"wsh(or_d(pk({primary_xpub}/<0;1>/*),and_v(v:pkh({recovery_xpub}/<0;1>/*),older({csv_value}))))"
     )
 
     lianad = Lianad(
         datadir,
-        owner_hd,
-        recovery_hd,
+        signer,
         main_desc,
         bitcoind.rpcport,
         bitcoind_cookie,
