+++ tests/test_framework/lianad.py
@@ -2,7 +2,6 @@ import logging
 import os
 import shutil
 
-from bip32.utils import coincurve
 from bip380.descriptors import Descriptor
 from bip380.miniscript import SatisfactionMaterial
 from test_framework.utils import (
@@ -15,11 +14,9 @@ from test_framework.utils import (
 )
 from test_framework.serializations import (
     PSBT,
-    sighash_all_witness,
     CTxInWitness,
     CScriptWitness,
     PSBT_IN_BIP32_DERIVATION,
-    PSBT_IN_WITNESS_SCRIPT,
     PSBT_IN_PARTIAL_SIG,
     PSBT_IN_FINAL_SCRIPTWITNESS,
 )
@@ -29,8 +26,7 @@ class Lianad(TailableProc):
     def __init__(
         self,
         datadir,
-        owner_hd,
-        recovery_hd,
+        signer,
         multi_desc,
         bitcoind_rpc_port,
         bitcoind_cookie_path,
@@ -40,8 +36,7 @@ class Lianad(TailableProc):
         self.datadir = datadir
         self.prefix = os.path.split(datadir)[-1]
 
-        self.owner_hd = owner_hd
-        self.recovery_hd = recovery_hd
+        self.signer = signer
         self.multi_desc = multi_desc
         self.receive_desc, self.change_desc = multi_desc.singlepath_descriptors()
 
@@ -64,51 +59,6 @@ class Lianad(TailableProc):
             f.write("[bitcoind_config]\n")
             f.write(f"cookie_path = '{bitcoind_cookie_path}'\n")
             f.write(f"addr = '127.0.0.1:{bitcoind_rpc_port}'\n")
-
-    def sign_psbt(self, psbt, recovery=False):
-        """Sign a transaction.
-
-        This will fill the 'partial_sigs' field of all inputs. Uses either the 'primary'
-        'recovery' key as specified.
-
-        :param psbt: PSBT of the transaction to be signed.
-        :returns: PSBT with a signature in each input for the owner's key.
-        """
-        assert isinstance(psbt, PSBT)
-
-        # Which key to sign the transaction with.
-        hd = self.recovery_hd if recovery else self.owner_hd
-
-        # Sign each input.
-        for i, psbt_in in enumerate(psbt.i):
-            # First, gather the needed information from the PSBT input.
-            # 'hd_keypaths' is of the form {pubkey: (fingerprint (4 bytes), derivation path (n * 4 bytes))}
-            fing_der = next(iter(psbt_in.map[PSBT_IN_BIP32_DERIVATION].values()))
-            raw_der_path = fing_der[4:]
-            der_path = [
-                int.from_bytes(raw_der_path[i : i + 4], byteorder="little", signed=True)
-                for i in range(0, len(raw_der_path), 4)
-            ]
-            script_code = psbt_in.map[PSBT_IN_WITNESS_SCRIPT]
-
-            # Now sign the transaction.
-            sighash = sighash_all_witness(script_code, psbt, i)
-            privkey = coincurve.PrivateKey(hd.get_privkey_from_path(der_path))
-            pubkey = privkey.public_key.format()
-            assert pubkey in psbt_in.map[PSBT_IN_BIP32_DERIVATION].keys(), (
-                der_path,
-                fing_der,
-                pubkey,
-                psbt_in.map[PSBT_IN_BIP32_DERIVATION].keys(),
-            )
-            sig = privkey.sign(sighash, hasher=None) + b"\x01"
-            logging.debug(
-                f"Adding signature {sig.hex()} for pubkey {pubkey.hex()} (path {der_path})"
-            )
-            assert PSBT_IN_PARTIAL_SIG not in psbt_in.map
-            psbt_in.map[PSBT_IN_PARTIAL_SIG] = {pubkey: sig}
-
-        return psbt
 
     def finalize_psbt(self, psbt):
         """Create a valid witness for all inputs in the PSBT.
